windows pt 3

+++ install_and_update.py
@@ -15,7 +15,8 @@ try:
         subprocess.run('git clone https://github.com/Azure-Fey-MC/Power-Generation.git'.split(" "))
         repo = os.listdir(repo_folder)
         for item in repo:
-            os.rename(repo_folder+item, item)
+            if os.path.exists(os.path.join(repo_folder, item) and operating_system == "Windows"):
+                os.rename(repo_folder+item, item)
         os.rmdir(repo_folder)
 
     import roll
